requirements.txt
discord
requests
pyaml
json
asyncio
